refactor(textinput): move TextInput to its folder with theme interface
#443 #234 #460
