+++ src/lib/components/FormControls/index.ts
@@ -1,4 +1,3 @@
 export * from './HelperText';
 export * from './Textarea';
-export * from './TextInput';
 export * from './ToggleSwitch';
