+++ src/lib/components/Modal/Modal.stories.tsx
@@ -3,8 +3,8 @@ import type { Meta, Story } from '@storybook/react/types-6-0';
 import { HiOutlineExclamationCircle } from 'react-icons/hi';
 import { Button } from '../Button';
 import { Checkbox } from '../Checkbox';
-import { TextInput } from '../FormControls';
 import { Label } from '../Label';
+import { TextInput } from '../TextInput';
 import type { ModalProps } from './Modal';
 import { Modal } from './Modal';
